+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - **Breaking change** `api.rpc.state.queryStorage(...)` now properly decodes the `Vec<StorageChangeSet>`. this means that results will come back as `[Hash, Codec[]][]` when using this RPC.
 - `StorageKey` now have an `.args` property, decoded from meta in the cases twox64concat is used on maps
-- Fix storage `.entries` type conversions to return exact types (not `Option` in some cases)
+- Fix `api.query.*.*.entries` type conversions to return exact types (not `Option` in some cases)
+- Add `api.query.*.*.keys` to retrieve only the storage keys, similar to `.entries`
 - Full linked map retrievals will now use direct getStorage queries for faster operation
 - Underlying rpc-core interfaces now unwraps `Error("...")` when found in responses
 - Added `derive.eras*` interfaces for queries to new Substrate staking interfaces